pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "prompt-injection-detector"
version = "1.0.0"
description = "Real-time NLP security middleware for detecting and intercepting adversarial prompt injection attacks"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.0",
    "transformers>=4.35",
    "datasets>=2.14",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "scikit-learn>=1.3",
    "numpy>=1.24",
]

[project.optional-dependencies]
onnx = [
    "optimum[onnxruntime]>=1.14",
    "onnxruntime>=1.16",
]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "httpx>=0.25",
    "ruff>=0.1",
    "pre-commit>=3.5",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
